+++ cookbooks/chub-sensu/attributes/default.rb
@@ -1,5 +1,0 @@
-default["sensu"]["use_ssl"] = false
-default['chub-sensu']['email']['sender'] = 'sensu@commercehub.com'
-default['chub-sensu']['email']['smtp_server'] = 'mail.commercehub.com'
-default['chub-sensu']['email']['smtp_port'] = 25
-default['chub-sensu']['email']['smtp_domain'] = 'commercehub.com'