+++ cookbooks/chub-sensu/roles/monitor.rb
@@ -1,2 +1,0 @@
-name "monitor"
-description "Used for sensu checks that only should run on one node"