+++ cookbooks/chub_sensu_sso/recipes/default.rb
@@ -1,0 +1,10 @@
+#
+# Cookbook Name:: chub_sensu_sso
+# Recipe:: default
+#
+# Copyright (C) 2014 CommerceHub
+# 
+# All rights reserved - Do Not Redistribute
+#
+
+Chef::Application.fatal!("Use chub_sensu_sso::server or chub_sensu_sso::client, not the default recipe")